+++ etna/analysis/outliers/confidence_interval_outliers.py
@@ -14,10 +14,34 @@ if TYPE_CHECKING:
     from etna.models import SARIMAXModel
 
 
+def create_ts_by_column(ts: "TSDataset", column: str) -> "TSDataset":
+    """Create TSDataset based on original ts with selecting only column in each segment and setting it to target.
+
+    Parameters
+    ----------
+    ts:
+        dataset with timeseries data
+    column:
+        column to select in each.
+
+    Returns
+    -------
+    result: TSDataset
+        dataset with selected column.
+    """
+    from etna.datasets import TSDataset
+
+    new_df = ts[:, :, [column]]
+    new_columns_tuples = [(x[0], "target") for x in new_df.columns.tolist()]
+    new_df.columns = pd.MultiIndex.from_tuples(new_columns_tuples, names=new_df.columns.names)
+    return TSDataset(new_df, freq=ts.freq)
+
+
 def get_anomalies_confidence_interval(
     ts: "TSDataset",
     model: Union[Type["ProphetModel"], Type["SARIMAXModel"]],
     interval_width: float = 0.95,
+    in_column: str = "target",
     **model_params,
 ) -> Dict[str, List[pd.Timestamp]]:
     """
@@ -27,27 +51,37 @@ def get_anomalies_confidence_interval(
     Parameters
     ----------
     ts:
-        TSDataset with timeseries data(should contains all the necessary features).
+        dataset with timeseries data(should contains all the necessary features).
     model:
-        Model for confidence interval estimation.
+        model for confidence interval estimation.
     interval_width:
-       The significance level for the confidence interval. By default a 95% confidence interval is taken.
+        the significance level for the confidence interval. By default a 95% confidence interval is taken.
+    in_column:
+        column to analyzes
+        If it is set to "target", then all data will be used for prediction.
+        Otherwise, only column data will be used.
 
     Returns
     -------
     dict of outliers: Dict[str, List[pd.Timestamp]]
-        Dict of outliers in format {segment: [outliers_timestamps]}.
+        dict of outliers in format {segment: [outliers_timestamps]}.
 
     Notes
     -----
-    Works only with target column.
+    For not "target" column only column data will be used for learning.
     """
+    if in_column == "target":
+        ts_inner = ts
+    else:
+        ts_inner = create_ts_by_column(ts, in_column)
     outliers_per_segment = {}
     time_points = np.array(ts.index.values)
-    model_instance = model(interval_width=interval_width, **model_params)
-    model_instance.fit(ts)
-    confidence_interval = model_instance.forecast(deepcopy(ts), confidence_interval=True)
-    for segment in ts.segments:
+    model_instance = model(**model_params)
+    model_instance.fit(ts_inner)
+    confidence_interval = model_instance.forecast(
+        deepcopy(ts_inner), confidence_interval=True, interval_width=interval_width
+    )
+    for segment in ts_inner.segments:
         segment_slice = confidence_interval[:, segment, :][segment]
         anomalies_mask = (segment_slice["target"] > segment_slice["target_upper"]) | (
             segment_slice["target"] < segment_slice["target_lower"]