+++ etna/models/nn/deepar.py
@@ -3,16 +3,19 @@ from typing import Optional
 from typing import Union
 
 import pandas as pd
-import pytorch_lightning as pl
-from pytorch_forecasting.data import TimeSeriesDataSet
-from pytorch_forecasting.models import DeepAR
-from pytorch_lightning import LightningModule
 
+from etna import SETTINGS
 from etna.datasets.tsdataset import TSDataset
 from etna.loggers import tslogger
 from etna.models.base import Model
 from etna.models.base import log_decorator
 from etna.transforms import PytorchForecastingTransform
+
+if SETTINGS.torch_required:
+    import pytorch_lightning as pl
+    from pytorch_forecasting.data import TimeSeriesDataSet
+    from pytorch_forecasting.models import DeepAR
+    from pytorch_lightning import LightningModule
 
 
 class DeepARModel(Model):